+++ src/Backend/Http/Controllers/Auth/ConfirmPasswordController.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Mymo\Backend\Http\Controllers\Auth;
+
+use App\Http\Controllers\Controller;
+use Mymo\Core\Traits\ResponseMessage;
+
+class ConfirmPasswordController extends Controller
+{
+    use ResponseMessage;
+}