+++ src/Backend/Http/Controllers/Backend/Setting/SocialLoginController.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Mymo\Backend\Http\Controllers\Backend\Setting;
+
+use Illuminate\Http\Request;
+use Mymo\Backend\Http\Controllers\Controller;
+
+class SocialLoginController extends Controller
+{
+
+}