+++ src/Backend/resources/lang/en/validation.php
@@ -1,0 +1,31 @@
+<?php
+
+return [
+    'attributes' => [
+        'email' => 'Email Address',
+        'name' => 'Name',
+        'password' => 'Password',
+        'title' => 'Title',
+        'description' => 'Description',
+        'body' => 'Body',
+        'status' => 'Status',
+        'thumbnail' => 'Thumbnail',
+        'categories' => 'Categories',
+        'theme' => 'Theme',
+        'anyone-can-register' => 'Anyone can register',
+        'user-confirmation' => 'User confirmation',
+        'email_host' => 'Host',
+        'email_port' => 'Port',
+        'email_encryption' => 'Encryption',
+        'email_username' => 'Username',
+        'email_password' => 'Password',
+        'email_from_address' => 'From address',
+        'email_from_name' => 'From name',
+        'email_setting' => 'Use above setting instead of env file setting',
+        'email_templates' => 'Email templates',
+        'password_confirmation' => 'Confirm password',
+        'avatar' => 'Avatar',
+        'language' => 'Language',
+        'site_url' => 'Site url',
+    ],
+];