+++ src/Core/Providers/MymoCoreServiceProvider.php
@@ -13,9 +13,9 @@
 namespace Mymo\Core\Providers;
 
 use Illuminate\Support\Facades\Schema;
+use Mymo\Backend\Providers\MymoBackendServiceProvider;
+use Mymo\Backend\Providers\RouteServiceProvider;
 use Mymo\Core\Helpers\HookAction;
-use Mymo\Core\Http\Middleware\Admin;
-use Mymo\Core\Macros\RouterMacros;
 use Mymo\Email\Providers\EmailTemplateServiceProvider;
 use Mymo\FileManager\Providers\FilemanagerServiceProvider;
 use Mymo\Notification\Providers\NotificationServiceProvider;
@@ -25,7 +25,6 @@ use Mymo\Repository\Providers\RepositoryServiceProvider;
 use Mymo\Theme\Providers\ThemeServiceProvider;
 use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Facades\Validator;
-use Illuminate\Routing\Router;
 use Mymo\Updater\UpdaterServiceProvider;
 use Mymo\Installer\Providers\InstallerServiceProvider;
 
@@ -34,12 +33,10 @@ class MymoCoreServiceProvider extends ServiceProvider
     public function boot()
     {
         $this->bootMigrations();
-        $this->bootMiddlewares();
         $this->bootPublishes();
         $this->loadMigrationsFrom(core_path('database/migrations'));
         $this->loadFactoriesFrom(core_path('database/factories'));
-        $this->loadViewsFrom(core_path('resources/views'), 'mymo_core');
-        $this->loadTranslationsFrom(core_path('resources/lang'), 'mymo_core');
+
 
         Validator::extend('recaptcha', 'Mymo\Core\Validators\Recaptcha@validate');
         Schema::defaultStringLength(150);
@@ -49,21 +46,15 @@ class MymoCoreServiceProvider extends ServiceProvider
     {
         $this->registerProviders();
         $this->registerSingleton();
-        $this->registerRouteMacros();
         $this->mergeConfigFrom(__DIR__ . '/../config/mymo_core.php', 'mymo_core');
     }
 
     protected function bootMigrations()
     {
-        $mainPath = __DIR__ . '/../database/migrations';
+        $mainPath = __DIR__ . '/../../../database/migrations';
         $directories = glob($mainPath . '/*' , GLOB_ONLYDIR);
         $paths = array_merge([$mainPath], $directories);
         $this->loadMigrationsFrom($paths);
-    }
-
-    protected function bootMiddlewares()
-    {
-        $this->app['router']->aliasMiddleware('admin', Admin::class);
     }
 
     protected function bootPublishes()
@@ -71,15 +62,12 @@ class MymoCoreServiceProvider extends ServiceProvider
         $this->publishes([
             __DIR__ . '/../config/mymo_core.php' => base_path('config/mymo_core.php'),
         ], 'mymo_config');
-
-        $this->publishes([
-            __DIR__.'/../resources/assets' => public_path('mymo/styles'),
-        ], 'mymo_assets');
     }
 
     protected function registerProviders()
     {
         //$this->app->register(UpdaterServiceProvider::class);
+        $this->app->register(MymoBackendServiceProvider::class);
         $this->app->register(InstallerServiceProvider::class);
         $this->app->register(DbConfigServiceProvider::class);
         $this->app->register(RouteServiceProvider::class);
@@ -99,10 +87,5 @@ class MymoCoreServiceProvider extends ServiceProvider
         $this->app->singleton('mymo.hook', function () {
             return new HookAction();
         });
-    }
-
-    protected function registerRouteMacros()
-    {
-        Router::mixin(new RouterMacros());
     }
 }